server/container_create: Simplify Kube/OCI process merge
Handling both the nil ociConfig check and kubeCommands length check on
the same line lets us drop one level of nesting.  And that initial
zero-length check makes the internal kubeCommands nil check redundant,
so we can drop that too.

Also adjust the buildOCIProcessArgs argument from a *v1.Image to
*v1.ImageConfig, because we don't need anything else from Image here.
This moves the nil Image check from inside buildOCIProcessArgs to the
calling function, but I think that's worth the tighter scoping.

Signed-off-by: W. Trevor King <wking@tremily.us>