Permanently disable swap prior to testing
When kubernetes is started with active swap, it shows:

``WARNING : The kubelet is configured to not fail if swap is enabled;
production deployments should disable swap.``

However, a swap configuration may be desirable for compiling/building
prior to testing.  Add a handful of ``post_tasks`` at the last moment
before testing, to turn off and permanently disable swap.

This allows swap to be configured/setup prior to this point if desired,
while guaranteeing it is not available after.

Signed-off-by: Chris Evich <cevich@redhat.com>