Add extra info to verbose requests to PodSandboxStatus
If PodSandboxStatusRequest.Verbose is true now we are returning the cri-o
version in a JSON object for debug purposes. In the future extra information
(to be defined) should be added to the response

In order to avoid problems when we execute the tests in parallel the fixtures
for new test sandbox and container are creating their own random IDs and
returning them in case you need to refer to them.

Finally, "make testunit" is being run as root to solve a problem with a `chown`
that couldn't be performed otherwise.

This commit closes #1144

Signed-off-by: Álex González <agonzalezro@gmail.com>