Get the terminal size of stdout and set it inside of container
podman needs to have the size of the tty of the terminal it is
running in passed into the tty that will be used by the container.

This seems to fix the issue we are seeing with podman, but not sure
if it breaks cri-o.  Would their be a stdout when CRI-O runs.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>