Add read-only flag
Add a read-only flag to the cri-o config.  This will force all containers
run on the system to run in readonly mode.  In order to make most containers
work this patch mounts a tmpfs on /run, /tmp and /var/tmp so that these directories can continue to be written to, but the rest of the file system will be
read-only.  If the container needs to write to other locations, then volumes
must be mounted into the container.

Running in read-only mode increases the container security, since a hacked
container will not be able to overwrite existing code.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>
Signed-off-by: umohnani8 <umohnani@redhat.com>