oci: Force kill the container process only if nothing else worked
The function waitContainerStop() in oci/oci.go was designed to check
on a container process and wait for it until it returns, or returns
an error if a timeout was reached.

Now, looking at the way this function is used, it is called twice
from the same function StopContainer() with the idea of trying to
wait the process properly first. Then, if the process is not
terminating as expected, another signal is sent to the container
process, and another call to waitContainerStop() is issued.

This second call should be the only one calling into SIGKILL because
this is the last chance to stop the container process. But from the
first call into this function, we should only retrieve sort of a
boolean value telling if the process is still around, instead of
trying to kill it with a SIGKILL already.

This commit modifies waitContainerStop() by adding a simple flag
indicating if this function should try to force kill the container
process using SIGKILL signal in case the process is still around
after the timeout has been reached.

Fixes #1327

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>