conmon: add support to restore a container
runc supports checkpointing and restoring containers with the help of
CRIU. To checkpoint a container from podman it is enough to just call
runc to checkpoint the container. To restore a container with podman the
resulting container should again be under the control of conmon.

This extends conmon to be able to also restore a container.

Signed-off-by: Adrian Reber <areber@redhat.com>