.travis.yml: Add allowed-to-fail OS X compilation
As an initial step in continuous integration testing for Vincent's
cross-platform work.  Once we get this passing on OS X we can remove
the allow_failures entry.  The best docs I could find for using
allow_failures with build stages was [1], so that's the pattern I'm
following.  Travis' OS X docs are in [2].

The TRAVIS_OS_NAME approach is recommended in [3].  We're not going to
actually use any of the storage drivers yet, so we don't need to
install btrfs, etc.  And the OS X workers already come with automake
[4].  We'll see how things work with the other requirements missing

[1]: https://github.com/travis-ci/travis-ci/issues/7789#issuecomment-303685412
[2]: https://docs.travis-ci.com/user/reference/osx
[3]: https://docs.travis-ci.com/user/installing-dependencies/#Installing-Dependencies-on-Multiple-Operating-Systems
[4]: https://docs.travis-ci.com/user/reference/osx#Compilers-and-Build-toolchain

Signed-off-by: W. Trevor King <wking@tremily.us>