.travis.yml: Only install OSTree on Linux
Avoid [1]:

  autoreconf: running: /usr/local/Cellar/autoconf/2.69/bin/autoconf --force
  configure.ac:78: error: possibly undefined macro: AM_PATH_GLIB_2_0
        If this token and others are legitimate, please use m4_pattern_allow.
        See the Autoconf documentation.

and likely a bunch of later errors masked by that one, because OSTree
is Linux-only.

The YAML syntax I'm using is a folded block style [2], and it allows
us to make the whole OSTree checkout-and-build a single operation.

I'm also using a subshell for the directory change (instead of
pushd/popd).  I'm not sure if the working directory perisist to the
next install step or not (the previous pushd approach suggests it
does).  But with a subshell we don't have to mess with the parent
working directory at all, so there's no need to clean up.

[1]: https://travis-ci.org/kubernetes-incubator/cri-o/jobs/350679059#L200
[2]: http://www.yaml.org/spec/1.2/spec.html#id2796251

Signed-off-by: W. Trevor King <wking@tremily.us>