hack/seccomp_tag: Stop hardcoding the seccomp build tag
We've been hardcoding it since 0e4af6d6 (*: add seccomp buildtag,
2016-11-29, #219), but that predates the hack/* approach which landed
in f893e38d (Add build tags for integration tests, 2016-12-07, #210).
I'm not sure why #210 skipped libseccomp (possibly because we always
install it on Linux in .travis.yml), but now that we are trying to
build on OS X too we need to be more flexible.

Signed-off-by: W. Trevor King <wking@tremily.us>