pkg/findprocess: Add a new package for portable process detection
The traditional POSIX approach to detecting processes is to kill(pid,
0) them and check for ESRCH [1].  However, Go's os.Process makes
detecting that a bit awkward.

On Unix, Go's os.FindProcess implementation always succeeds [2].  And
os.Process's Signal translates ESRCH [3] to an internal errFinished
[4].  The Unix wrapper here rolls the FindProcess and Signal(0) calls
together, and translates os's errFinished to a public ErrNotFound
(because gometalinter complains about ESRCH).

On Windows, os.Process's Signal only supports kills [5], so we can't
perform a check like POSIX's kill(pid, 0) [1].  Instead, this commit
assumes os.FindProcess will return errors on nonexistent processes,
although that behavior is not covered in the Go docs [2].  The Windows
os.FindProcess wraps an OpenProcess call [6].  OpenProcess can set a
few errors, including ERROR_INVALID_PARAMETER and ERROR_ACCESS_DENIED
[7].  It's possible there's an ESRCH analog in their system error
codes [8], but I'm not going to dig through and find it ;).  Once
someone can test this on a Windows machine, they can tell us what
error(s) to look for.

Another possible portability issue is the use of 'sleep' in the tests.
That's fine for POSIX systems [9], but we may need to slot in another
command for Windows.

Use the new package for portable process-existence checks in the oci
package.

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/functions/kill.html
[2]: https://golang.org/pkg/os/#FindProcess
     > On Unix systems, FindProcess always succeeds and returns a
     > Process for the given pid, regardless of whether the process
     > exists.
[3]: https://github.com/golang/go/blob/010579c2377654e575d46857ec7dc77bab586438/src/os/exec_unix.go#L72-L73
[4]: https://github.com/golang/go/blob/010579c2377654e575d46857ec7dc77bab586438/src/os/exec_unix.go#L53
[5]: https://github.com/golang/go/blob/010579c2377654e575d46857ec7dc77bab586438/src/os/exec_windows.go#L65-L71
[6]: https://github.com/golang/go/blob/010579c2377654e575d46857ec7dc77bab586438/src/os/exec_windows.go#L92-L95
[7]: https://msdn.microsoft.com/en-us/library/windows/desktop/ms684320(v=vs.85).aspx
[8]: https://msdn.microsoft.com/en-us/library/windows/desktop/ms681381(v=vs.85).aspx
[9]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/sleep.html

Signed-off-by: W. Trevor King <wking@tremily.us>