Fix accidental positive side-effect of RHEL 7.4- bug
There is a bug-fix in 7.5.0 which fixes containers access to cgroups on
the host, despite the default container_manage_cgroup == 0.  When
testing with 7.5.0 and later, this access will be blocked unless this
boolean is enabled.  Since it's also required in Fedora, and there's
no negative consequence for using it in <= 7.4, simply enable it for all
VMs.

Signed-off-by: Chris Evich <cevich@redhat.com>