Default to the storage defaults
Change the config to use github.com/containers/storage by default.
Allow users to override in the config, but if the admin does not
override the config.

Since containers/storage default driver is overlay, I removed all
reference to overlay2,  container/storage uses the same driver for
overlay and overlay2.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>