*: Return err from AddHook and ok from RemoveHook
Also rename errNotJSON to the public ErrNoJSONSuffix.  This allows us
to push success/failure logging out into server/server.go which
understands that it is not filtering for .json suffixes.  With this
change, we no longer log noise when someone:

* Touches a non-.json path in the watched directories.
* Removes a non-hook path in the watched directories.

Addition now only results in a logged success or error; previously
we'd log the "adding" attempt and possibly *also* log the hook-reading
error.

Signed-off-by: W. Trevor King <wking@tremily.us>