conmon/pause: Allow user-specified CFLAGS
-std, -O, and -W are all options that the caller should be able to set
for themselves.  My personal preference is to drop these defaults (and
leave it up to the compiler and system configuration to get reasonable
defaults [1]), but that approach was not acceptable to the CRI-O
maintainers [2].  This commit moves options that CRI-O doesn't really
need into a fallback ?= definition [3].  The options that we do need
stay in the 'override +=' append [4].  The main drawback to this
approach (vs. dropping the ?= default completely) is that a user
setting, for example, CFLAGS=-O2 will clobber our default -std,
etc. values.  But I can't see a way around that besides the approach
in [1].

[1]: https://github.com/kubernetes-incubator/cri-o/pull/1361
[2]: https://github.com/kubernetes-incubator/cri-o/pull/1361#issuecomment-373504858
[3]: https://www.gnu.org/software/make/manual/html_node/Setting.html
[4]: https://www.gnu.org/software/make/manual/html_node/Override-Directive.html

Signed-off-by: W. Trevor King <wking@tremily.us>