WIP: use ID mappings features in storage
Use ID mapping features in the storage library to configure sandboxes
and containers to use user namespaces that match our configuration.

If a sandbox or container that we're creating is privileged, force use
of the host/node user namespace.

For unprivileged pod sandboxes, if the RunPodSandboxRequest contains
annotations providing UID or GID mappings to be used, or we have default
mappings configured for the storage library or set via our configuration
file or on the command line, put the pod sandbox in its own user
namespace.

For unprivileged containers, if the CreateContainerRequest does not
contain UID or GID mapping annotations, use the pod's sandbox's user
namespace and mappings.  Otherwise, use the supplied UID and GID
mappings, or defaults if one of them is not specified, and put the
container into its own user namespace.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>