oci: Replace 'c.ID()' with 'c.id'
We don't need to use the public getter from this same-package code.
This may save the cost of a function call (or maybe the Go compiler
optimizes that out).  This exposes us to more maintenance work if we
ever rename Container.id, but I don't expect us to do that.  And it
makes us consistent with the many more c.id calls which are already in
oci.go.

Signed-off-by: W. Trevor King <wking@tremily.us>