test: Fix race condition in ctr.bats
When running BATS testing from ctr.bats, it works very well for
runc, but in case of Kata Containers, we can expect actions to
take longer time to be achieved. In the specific case of the test
"ctr /etc/resolv.conf rw/ro mode" from ctr.bats, we should not expect
the container state to be CONTAINER_EXITED right after the start
returned. Indeed, it might take some time for the container to exit
and the call to "crictl inspect $ctr_id --output table" should be
tried several times before we consider the output as failing.

This patch relies on the "ctr oom" test as an example of how to
implement this retry behavior.

This is a backport of #1489

Fixes #1491

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>