container_create: simplify missing src directory creation
Attempt to resolve the symlink before attempting to create a missing
source directory.  In this way we can re-use the error code we got
from the Lstat in resolveSymbolicLink without repeating it a second
time.

As part of the commit, change the error handling for os.MkdirAll to
report the correct error (err1).

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>