lib/container_server: Allow the same command in multiple hook files
Lift the constraint added in ca94095 (server: fsnotify on hooks,
2018-02-16, #1345).  For example, you may want to limit a hook to one
stage for an annotations match and another stage for a cmd match.
Using two files with the same hook entry is the only way to do that.

This is very similar to dbd16413 (lib/hooks: Allow the same command in
multiple hook files, 2018-03-26, #1468), where I'd attempted to remove
the constraint from ca94095, but actually ended up removing a very
similar constraint which had been added in 139d0841 (Add support for
oci-hooks to libkpod, 2017-08-12, #562).

Signed-off-by: W. Trevor King <wking@tremily.us>