conmon: return the correct exit status
WEXITSTATUS(x) can be used only if WIFEXITED(x).  We were using it
also in the case the process exited because it was signaled.

Since we check for the exit_status both for the container and the
runtime, move the common code to a new function get_exit_status that
is used in both cases.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>