pause: remove pause binary from repo
This code was added back when cri-o didn't have the ability to pull
images, and the purpose of 1313f0dd720f ("*: add pause binary as a build
target") was to make it easier to play with cri-o without Docker.

However, cri-o has long-since grown support for image pulling and the
code that used to use the pre-compiled 'pause' binary now uses the
'pause' image. Keeping this binary around doesn't make sense anymore,
especially since it was originally just a copy of the upstream
Kubernetes one.

Signed-off-by: Aleksa Sarai <asarai@suse.de>