network: Fix manage NetworkNS lifecycle
After port ManageNetworkNSLifecycle from 1.9 branch the logic
was modified to not save the host network state from the sandbox.

This commit restore the logic to save if the sandbox host network. It saves
if the sandbox is using network host-namespace. Then when a container is
created LoadSandbox will check for the annotation to know if the sandbox
where the container is created was using host NetworkNS.
was modified to not save the hostnetwork state from the sandbox.

Fixes: #1523

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>