server/sandbox: clean up network even if networkStart() fails
If the CNI network plugin completes successfully, but something fails
between that success and CRIO's sandbox setup code, plugin resources
may not be cleaned up.  A non-trivial amount of code runs after the
plugin itself exits and ocicni's SetUpPod() returns, and any error
condition recognized by that code would cause this leakage.

The Kubernetes CRI RunPodSandbox() request does not attempt to clean
up on errors, since it cannot know how much (if any) networking
was actually set up.  It depends on the CRI implementation to do
that cleanup for it.

Concrete examples include if the sandbox's container is somehow
removed during during that time, or another OS error is encountered,
or the plugin returns a malformed result to ocicni.

ref: https://bugzilla.redhat.com/show_bug.cgi?id=1532965

Signed-off-by: Dan Williams <dcbw@redhat.com>