Update vendor for containers/image and containers/storage
container/storage now supports devicemapper options that allow you to
configure it.  This might be useful to katacontainers.

Containers/image has had a few bugfixes lately and we can get the benefit of these.
Containers/image has grown better support for User Namespace, which will allow us to
run containers with both a UserMapping as well as the hostmapping, and have the processes
see the correct Ownership according to the current mapping.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>