Allow running int. tests with userns enabled
Added a new ``userns`` tag, which enables running the integration tests
(``test.yml``) with usernamespaces.  This is primarily affected by
setting the ``TEST_USERNS=1`` env. var.  However, due to increasing
complexity of the command, it's not separated from it's env. vars
in the ``vars.yml`` file.

An additional task was added, contingent on the ``userns`` tag, which
updates the environment variables as required for usernamespace testing.
This allows both sets of env. vars, and the command to all be defined in
a single, easy to find place.

Signed-off-by: Chris Evich <cevich@redhat.com>