Simplify definition of STORAGE_OPTIONS
Normally the integration tests execute with
``--storage-driver=overlay``.  However, when testing either RHEL or
CentOS, ``overlay.override_kernel_check=1`` is also needed.
Previously this was done by secondary mangling of the testing
command-line contents.

Simplify this to meet POLA standards, by templating the value of the
``STORAGE_OPTIONS`` environment variable.  This allows the value
to be defined once, with clear content-logic, and without the
for mangling.

Signed-off-by: Chris Evich <cevich@redhat.com>