WIP - Proposed Stream Log Writer
This is a conceptual PR, meant to illustrate a proposed stream log file
format that removes the byte level interpretation of stdout/stderr in
favor of simply recording what data was read on each system call. It
defers the interpretation of the byte stream to the consumer, allowing
this writer to operate with as little overhead as possible (avoiding bad
containers that write only newlines, or small numbers of bytes between
newlines). The reader of the byte stream is then tasked with
reassembling the stream according to whatever interpretation they see
fit to use.

The goal of this work is to provide a simple format that will stream
well into an object store, such that, given enough metadata stored with
the stream, the consumer can reconstruct the I/O stream at the time it
is read.

Signed-off-by: peter.portante@redhat.com