container_create: correctly relabel mounts when asked
We were wrongly setting "shared" to the label.Relabel call when
relabeling mounts.
Basically, we need to pass the equivalent to ":Z" in docker when
setting up and labeling mounts. That's provided by the CRI
"SelinuxRelabel" field. However, even if we checked that field, we were
wrongly setting shared (":z") which automatically sets label to just "s0".
This patch fixes the above by using shared=false when relabeling so
that a full Chcon is performed with the full label.

Signed-off-by: Antonio Murdaca <runcom@redhat.com>