Reduce amount of logs being printed by default
Changed the Info logs to Debug logs so that it is only
logged if an error occurs or if the log-level is set to debug.
In conmon, only log container status if it is non-zero and removed
the "about to waitpid" logs.
Also only drain stdout and stderr if a timeout does not occur.

Signed-off-by: umohnani8 <umohnani@redhat.com>