oci: Define a timeout for WaitContainerStateStopped()
A minimal timeout is required when waiting for the container
termination since runtimes like Kata Containers might take more
time than the standard runc.

This commit introduces a minimal timeout value that will ensure
a proper amount of time will be waited before to issue a timeout.

It also adds the possibility for the user to configure through the
configuration file crio.conf, any timeout that he would like to see
applied. The timeout defined through the configuration cannot be
less than the minimal of 10s. Any value inferior to 10 seconds will
simply be ignored.

Fixes #1695

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>