oci: update privilege/trust handling
Prior to this patch, the trusted runtime (ie, runc) was used if the
container was privileged, even if the workload was explicitly marked
untrusted and an untrusted runtime (kata, for example) was provided.

This patch will follow the explicit annotation, since there are cases where
privileged container is desired with kata containers.

When the default trust level is untrusted, privileged containers will still
make use of the default runtime unless the untrusted annotation is provided
for the workload. This is necessary to make sure core kube-system privileged
containers continue to function properly.

Fixes: #1662

Signed-off-by: Eric Ernst <eric.ernst@intel.com>