apparmor: use github.com/containers/libpod/pkg/apparmor
Use github.com/containers/libpod/pkg/apparmor and remove CRI-O's
pkg/apparmor.  This change is part of the gradual port to libpod and
ultimately reduces redundancy among container runtimes.  The new
apparmor package is also faster as it doesn't create a temp file when
loading a profile but passes it through a pipe.

The default apparmor profile "crio-default" is now suffixed with a dash
followed by version.Version and loaded unconditionally at server start.
This way, we assure smooth upgrades.

To catch future build regressions early, compile CRI-O with apparmor
support if libapparmor is installed on the host.

Notice that this change required to re-vendor libpod from its new home
github.com/containers/libpod and to change all references to it in the
code and the documentation and build files.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>