Add missing argument to logrus.Debugf() format string
Calls to logrus.Debugf() in lib/container.go are missing an argument in the
format string.

Go 1.11 catches the error, previous versions apparently allowed it to compile.

There is a fix on master in commit 0bd3087 that is part of a larger commit.
This commit backports the change to fix compilation with go 1.11 to branch
release-1.11.

Signed-off-by: Jeff Kowalczyk <jkowalczyk@suse.com>