main: allow defining default conf file path at buildtime
Some implementations require configuration files to be located
at custom directories.
This change allows defining in build time a custom path for the
distro-default crio.conf file through the variable DefaultsPath,
so that it is included in the search for a default config file.
For example:

go build -ldflags "-X main.DefaultsPath=$DEFAULT_CONFIG_PATH"

The search is made only in the usual locations if DefaultsPath
is absent.

Co-authored-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>
Co-authored-by: Jose Lamego <jose.a.lamego@intel.com>

Signed-off-by: Jose Lamego <jose.a.lamego@intel.com>