config: fix specified settings
I've noticed the cri-o RPM generates the configuration using this
snippet:

./bin/crio ... --storage-driver=overlay
--storage-opt="overlay.override_kernel_check=1" config ...

However, the storage options aren't actually set because they are
commented out in the config stub we use to generate the config file.
This has worked till now because /etc/containers/storage.conf contains
valid settings used implicitely by cri-o but we should not rely on that
since we're actually generating the configuration ourselves.

This commit add checks around the config generation to make sure the
correct options are picked and printed.
This commit also de-comment ulimits settings as that's the correct
thing to do since the empty list is the same as commenting it out and
also, it's needed if someone specifies a config with ulimits.

Signed-off-by: Antonio Murdaca <runcom@redhat.com>