conmon: fix case when symlink == max UNIX socket path
address a corner case where the symlink path length can end up to be
the same as the socket.  When it happens, the symlink prevents the
socket to be be created.  This could still be a problem with other
containers, but it is safe to assume the CUUIDs don't change length in
the same directory.  As a workaround, in such case, make the symlink
one char shorter.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>