Add additonal device support to CRI-O
Add support to CRI-O to have additional devices.  This will allow machines
with specialized hardware to have all of its containers gain access to the device.

For example if you have a nvidia machine with a special GPU, you can tell CRI-O
to pull the device into every one of its containers.

Signed-off-by: Qi Wang <qiwan@redhat.com>