lib: Move exec implementation to oci.Runtime
Exec is directly using the container runtime path, but this should be
handled from oci.Runtime directly instead.

Because we want to better support VM based container runtimes, it is
necessary to make sure no calls to the runtime binary are performed
from high layers such as lib package.

Fixes #1946

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>