lib: Move kill implementation to oci.Runtime
Kill is directly dealing with the runtime path, and this should be
specific to the way the oci.Runtime is actually implemented.

Because we want to better support VM based container runtimes, it is
necessary to make sure we don't allow code from lib package to access
directly the runtime binary path.

Fixes #1946

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>