lib: Move attach implementation to oci.Runtime
Attach is implemented fully assuming that conmon will listen to a
"ctl" file for terminal resizing, and also that conmon will create
an "attach" socket to connect to the IO of an existing container.

Because we want to better support VM based container runtimes, it is
necessary to make sure we don't make such assumptions from high layers
such as lib package. Instead, those implementation details should be
handled by the oci.Runtime directly.

Fixes #1946

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>