server: serialize StopPodSandbox for the same sandbox
StopPodSandbox can be called on the same sandbox concurrently multiple
times.  This is racy as there are more phases involved in stopping the
sandbox and they are not idempotent.  Prevent it with a lock for
sandbox that is used only by StopPodSandbox.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>
(cherry picked from commit 6414bad641b0d02bbc68acef3f7c850c633dd733)