oci: Abstract Runtime structure as a Go interface
This is the second step in making a clear separation between different
implementations of the runtime. This commit takes care of defining a
new Go interface RuntimeImpl that can be implemented depending on the
runtime version.

One technical detail is that we had to maintain Runtime as a Go
structure embedding this new Go interface, since there are a few
fields that are common to every implementations. Those common
fields are gathered beneath the new structure RuntimeBase.

Once the abstraction is clearly defined, this patch reuses the
former runtime implementation as the RuntimeV1 implementation.

Fixes #1989

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>